+++ src/main/java/com/book/management/dto/response/BookResponse.java
@@ -3,11 +3,13 @@ package com.book.management.dto.response;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 @Builder
 public class BookResponse {
   private Integer bookId;
-  private Integer isbn;
+  private Long isbn;
   private String bookTitle;
   private String bookAuthor;
 }
